refactor(effect): split state.cpp by concern, and close three audit items
state.cpp was 1410 lines, past the 1150 ceiling, and held four unrelated
concerns. Three move out, each to a file named for what unites it rather
than for where it came from:

- windowedfullscreen.cpp (327): monocle and windowed-fullscreen ownership.
  These share an invariant, which is why they travel together: both hold a
  ledger of a compositor state PlasmaZones imposed and only PlasmaZones can
  hand back, so membership sheds only on an arm that actually restores.
- pretilegeometry.cpp (253): the memory of a window's size before the engine
  touched it, filed under the bucket screen it was captured on.
- floatcleanup.cpp (329): the eligibility gate on the way in, the two float
  paths on the way out, and the membership they maintain.

state.cpp keeps (608) what the daemon publishes and the effect consumes:
the managed and scrolling screen sets, active layouts, focus-follows-mouse
and wheel focus, and the shortcut registration those drive. All per-session,
all dropped on daemon loss, all arriving on independent signals that can
transiently disagree — which is the hazard that unites them.

Pure move: identical symbol sets before and after, 34 definitions either
way, and the moved bodies diff byte-identical against their origin. Only
the headers are new, with includes trimmed per TU rather than copied
wholesale.

Verified in a NON-UNITY build as well, which is what a new TU needs and
what caught the one real bug here: screenschanged.cpp uses
PhosphorIdentity::WindowId with no include for it, and has only ever
compiled because unity batching placed it beside a TU that had one. That
predates this branch; four audit passes missed it because they read source
rather than compile lines.

Two audit items close alongside:

F18 — connectDaemonSubscriptions also wired existing windows, seeded the
cursor output and installed the input filter, none of it daemon-facing.
Those move to initExistingWindowsAndInput, called straight after it because
the sweep still has to follow the subscriptions.

G16 — setupWindowConnections had no idempotency guard, so a second call for
one window would double every per-window handler, and Qt::UniqueConnection
is illegal with the lambda slots it uses. It was safe only because the two
callers' window sets are disjoint, which was an argument rather than an
enforced rule. Now guarded, and erased in the windowDeleted backstop for
address-reuse safety.

I7 — the park reap skipped any window whose decoration slot was ceded to a
live shader transition, so a column parked in that state never got stamped
and the reap timer stopped, pinning its full-canvas targets for the session.
A ceded window now reaches the park half while still being held back from
the backdrop driver, which does need ownership. Two qCDebug lines make the
reap decision observable, because it is otherwise invisible from outside.

Build green in both unity and non-unity trees, 358/358 tests pass. The
modified reap loop is not live-exercised: the nested harness has no shader
or decoration packs assigned, so m_windowDecorations stays empty there and
the loop never runs.
